Refactor code, add comment-docs
