+++ Server/config.py
@@ -1,0 +1,15 @@
+from fastapi.security import OAuth2PasswordBearer
+from passlib.context import CryptContext
+
+# to get a string like this run:
+# openssl rand -hex 32
+SECRET_KEY = "09d25e094faa6ca2556c818166b7a9563b93f7099f6f0f4caa6cf63b88e8d3e7"
+ALGORITHM = "HS256"
+ACCESS_TOKEN_EXPIRE_MINUTES = 300
+
+keys = ["max_score", "avg_accuracy", "level", "max_speed_accuracy",
+        "time_spend", "last_visit", "max_symbols_per_day"]
+
+pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")
+
+oauth2_scheme = OAuth2PasswordBearer(tokenUrl="token")
