+++ main.py
@@ -1,155 +1,6 @@
-from datetime import timedelta
-from typing import Annotated
-
-import aiofiles
-from fastapi import Depends, FastAPI, HTTPException, status
-from fastapi.responses import FileResponse
-from fastapi.security import OAuth2PasswordBearer, OAuth2PasswordRequestForm
-from jose import JWTError, jwt
-from passlib.context import CryptContext
-from database.db import *
-
-from db.data_classes import *
-
-# to get a string like this run:
-# openssl rand -hex 32
-SECRET_KEY = "09d25e094faa6ca2556c818166b7a9563b93f7099f6f0f4caa6cf63b88e8d3e7"
-ALGORITHM = "HS256"
-ACCESS_TOKEN_EXPIRE_MINUTES = 300
-
-keys = ["max_score", "avg_accuracy", "level", "max_speed_accuracy",
-        "time_spend", "last_visit", "max_symbols_per_day"]
-
-pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")
-
-oauth2_scheme = OAuth2PasswordBearer(tokenUrl="token")
+from fastapi import FastAPI
+from Server.route import router
 
 app = FastAPI()
 
-
-def verify_password(plain_password, hashed_password):
-    return pwd_context.verify(plain_password, hashed_password)
-
-
-def get_password_hash(password):
-    return pwd_context.hash(password)
-
-
-def get_user(username: str):
-    user = get_person_by_username(username=username)
-    if user is not None:
-        return UserInDB(**user)
-
-
-def update_user_achievements(username: str, user: User):
-    set_achieve(data=user.achievements.dict(), name=user)
-
-
-def authenticate_user(username: str, password: str):
-    user = get_user(username)
-    if not user:
-        return False
-    if not verify_password(password, user.hashed_password):
-        return False
-    return user
-
-
-def create_access_token(data: dict, expires_delta: timedelta | None = None):
-    to_encode = data.copy()
-    if expires_delta:
-        expire = datetime.utcnow() + expires_delta
-    else:
-        expire = datetime.utcnow() + timedelta(minutes=15)
-    to_encode.update({"exp": expire})
-    encoded_jwt = jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)
-    return encoded_jwt
-
-
-async def get_current_user(token: Annotated[str, Depends(oauth2_scheme)]):
-    credentials_exception = HTTPException(
-        status_code=status.HTTP_401_UNAUTHORIZED,
-        detail="Could not validate credentials",
-        headers={"WWW-Authenticate": "Bearer"},
-    )
-    try:
-        payload = jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
-        username: str = payload.get("sub")
-        if username is None:
-            raise credentials_exception
-        token_data = TokenData(username=username)
-    except JWTError:
-        raise credentials_exception
-    user = get_user(username=token_data.username)
-    if user is None:
-        raise credentials_exception
-    return user
-
-
-async def get_current_active_user(
-        current_user: Annotated[User, Depends(get_current_user)]
-):
-    if current_user.disabled:
-        raise HTTPException(status_code=400, detail="Inactive user")
-    return current_user
-
-
-@app.post("/token", response_model=Token)
-async def login_for_access_token(
-        form_data: Annotated[OAuth2PasswordRequestForm, Depends()]
-):
-    user = authenticate_user(form_data.username, form_data.password)
-    if not user:
-        raise HTTPException(
-            status_code=status.HTTP_401_UNAUTHORIZED,
-            detail="Incorrect username or password",
-            headers={"WWW-Authenticate": "Bearer"},
-        )
-    access_token_expires = timedelta(minutes=ACCESS_TOKEN_EXPIRE_MINUTES)
-    access_token = create_access_token(
-        data={"sub": user.username}, expires_delta=access_token_expires
-    )
-    return {"access_token": access_token, "token_type": "bearer"}
-
-
-@app.get("/users/me/", response_model=User)
-async def read_users_me(
-        current_user: Annotated[User, Depends(get_current_active_user)]
-):
-    return current_user.dict()
-
-
-@app.post("/users/me/upload")
-async def upload_own_data(
-        current_user: Annotated[User, Depends(get_current_active_user)], user: User
-):
-    if get_person_by_username(username=user.username) is None:
-        raise HTTPException(
-            status_code=status.HTTP_400_BAD_REQUEST,
-            detail="Username already registered",
-        )
-    update_user_achievements(username=current_user.username, user=user)
-    return {"Success": True}
-
-
-@app.post("/signup", response_model=User)
-async def signup(user: User, password: str):
-    if get_person_by_username(username=user.username) is not None:
-        raise HTTPException(
-            status_code=status.HTTP_400_BAD_REQUEST,
-            detail="Username already registered",
-        )
-    hashed_password = get_password_hash(password)
-    user_to_pass = user.copy()
-
-    add_person(username=user.username, password=hashed_password, mail=user.email)
-
-    return user_to_pass
-
-
-@app.get("/files/words/{language}")
-async def send_words(language: Language):
-    if language in Language.en:
-        return FileResponse("Server/words/en.json", media_type="application/json", filename="en.json")
-    if language is Language.ru:
-        pass
-    raise HTTPException(status_code=400, detail="Invalid language")
+app.include_router(router)
